// app/interface.tsx
export interface FormData {
  fullName: string;
  email: string;
  phone: string;
  experience: string;
  currentPosition: string;
  expectedSalary: string;
  coverLetter: string;
  portfolio: string;
  cv: File | null;
}

export type FormCareerErrors = Partial<Record<keyof FormData, string>>;

export interface ApplicationFormSectionProps {
  selectedPosition: Pick<
    JobPosition,
    'title' | 'department' | 'location' | 'type' | 'experience' | 'deadline'
  >;
  formData: FormData;
  errors: FormCareerErrors;
  isSubmitting: boolean;
  handleInputChange: (field: keyof FormData, value: string) => void;
  handleFileChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
  handleSubmit: (e: React.FormEvent<HTMLFormElement>) => void;
  setIsOpen: (open: boolean) => void;
}

// ========== Job Postings ==========
export interface JobData {
  title: string;
  department: string;
  location: string;
  type: string;
  experience: string;
  featured?: boolean;
  description: string;
  requirements: string[];
  salary: string;
  deadline: Date;
}

export interface JobPosition extends JobData {
  gradient: string;
  icon: React.ElementType;
}